+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+### Version v0.1.1
+- Fix windows build bat
+- Catch OpenGL errors
+
 ### Version v0.1.0
 - Initial commit